% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracdet.R
\name{as.wd}
\alias{as.wd}
\alias{as.wd.Fracdet}
\title{Coerce to a wd object}
\usage{
as.wd(x)

\method{as.wd}{Fracdet}(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A \code{wd} object (see \code{\link[wavethresh]{wd}}).
}
\description{
Coarce an object to a \code{wd} object if possible.
}

