% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracdet.R
\name{coef.Fracdet}
\alias{coef.Fracdet}
\title{Extract fractional Brownian motion parameters}
\usage{
\method{coef}{Fracdet}(object, ...)
}
\arguments{
\item{object}{A \code{Fracdet} object.}

\item{...}{Ignored (used for S3 generic/method consistency).}
}
\value{
Named numeric vector with the least-squared
estimates of the fractional Brownian motion parameters
(\code{H} and \code{sigma2}).
}
\description{
Extract fractional Brownian motion parameters
}
\seealso{
\code{\link{Fracdet}}
}

