% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbmParameterEstimation.R
\name{estimatefBmPars}
\alias{estimatefBmPars}
\title{Estimate fractional Brownian motion parameters from its wavelet coefficients' variances}
\usage{
estimatefBmPars(x, use_resolution_levels, start = NULL, lower = NULL,
  upper = NULL, use_weights = FALSE, algorithm = "port", ...)
}
\arguments{
\item{x}{A \code{WaveletVar} object.}

\item{use_resolution_levels}{A numeric vector specifying which resolution levels
should be used in the fit.}

\item{start}{A named list or named numeric vector of starting estimates. The
names of the list should be \code{H} and \code{sigma2}.}

\item{lower, upper}{Vectors of lower and upper bounds. Note that the \code{H}
parameter should fulfil \eqn{0 < H < 1}, whereas that \code{sigma2} should be
\eqn{sigma2 > 0} If unspecified, proper bounds are computed.}

\item{use_weights}{Logical value indicating wheter the objective function is
weighted least squares or not. If \code{use_weights} is \code{TRUE}, proper
weights are automatically computed.}

\item{algorithm}{Character string specifying the algorithm to use (see
\code{\link[stats]{nls}}. Note that bounds can only be used with the "port"
algorithm (default choice).}

\item{...}{Additional \code{nls} arguments (see \code{\link[stats]{nls}}).}
}
\value{
A \code{nls} object representing the fitted model. See
\code{\link[stats]{nls}} for further details.
}
\description{
\code{estimatefBmPars} determines the fractional Brownian motion (fBm)
parameters by fitting the theoretical wavelet coefficients' variances
to given experimental variances. The estimation procedure is thus based
on nonlinear least-squares estimates and makes use of the \code{nls} function
\code{\link[stats]{nls}}.
}
\section{Warning}{

Do not use nls on artificial "zero-residual" data. See
\code{\link[stats]{nls}} for further details.
}
\examples{
set.seed(10)
fbm = fbmSim(n = 2 ^ 13, H = 0.4)
vpr = WaveletVar(wd(fbm, bc = "symmetric"))
plot(vpr)
# Estimate the fBm parameters using the largest resolution levels
# since the estimates of their variances are better
model = estimatefBmPars(vpr, use_resolution_levels = 5:12)
# The nls-fit is performed in semilog-space. Thus, a transformation
# of the predicted values is required
points(resolutionLevels(vpr),
       2 ^ predict(model, newdata = data.frame(x = 0:12)),
       col = 2,
       pch = 2)
# Since the estimates of the largest resolution levels are better we may
# use a weigthed regression scheme
wmodel = estimatefBmPars(vpr, use_resolution_levels = 5:12,
                         use_weights = TRUE)
points(resolutionLevels(vpr),
       2 ^ predict(wmodel, newdata = data.frame(x = 0:12)),
       col = 3,
       pch = 3)
legend("topright", pch = 1:3, col = 1:3, bty = "n",
       legend = c("Original data", "nls model", "weighted-nls model"))
}
\seealso{
\code{\link{WaveletVar}}, \code{\link{theoreticalWaveletVar}}
}

