% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbmSimulation.R
\name{fbmSim}
\alias{fbmSim}
\title{Simulate a discrete fractional Brownian motion}
\usage{
fbmSim(n, H)
}
\arguments{
\item{n}{Length of the time series.}

\item{H}{Hurst exponent.}
}
\value{
A numeric vector containing the simulated time series.
}
\description{
\code{fbmSim} simulates a fractional Brownian motion using the Levinson's
method.
}
\note{
The code of the \code{fbmSim} is largelly based on the implementation
 provided in the Coeurjolly article (see references).
}
\examples{
# simulate a dfbm process with H=0.2 and plot it
z = fbmSim(100, 0.2)
ts.plot(z)
}
\references{
Jean-Francois, Coeurjolly. "Simulation and Identification of the
Fractional Brownian Motion: A bibliographical and comparative study." (2007).
}

