% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbmSimulation.R
\name{fgnSim}
\alias{fgnSim}
\title{Simulate a discrete fractional Gaussian noise}
\usage{
fgnSim(n, H)
}
\arguments{
\item{n}{Length of the time series.}

\item{H}{Hurst exponent.}
}
\value{
A numeric vector containing the simulated time series.
}
\description{
Simulate a discrete fractional Gaussian noise
}
\details{
The FFT is used during the simulation so it is most efficient if
n is a power of 2.
}
\note{
The \code{fgnSim} is just a wrapper of the \code{SimulateFGN} function
from the \code{FGN} package.
}
\examples{
#simulate a dfgn process with H=0.2 and plot it
z = fgnSim(100, 0.2)
ts.plot(z)
}
\author{
A.I. McLeod
}
\seealso{
\code{\link[FGN]{SimulateFGN}}
}

