% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracdet.R
\name{getWaveletVar}
\alias{getWaveletVar}
\title{Get the wavelet coefficients' variances}
\usage{
getWaveletVar(x)
}
\arguments{
\item{x}{A \code{fracdet} object.}
}
\value{
A \code{waveletVar} object representing the wavelet coefficients'
variances of the wavelet transform represented by the \code{fracdet} object.
}
\description{
Get the wavelet coefficients' variances
}
\seealso{
\code{\link{fracdet}}
}

