% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracdet.R
\name{getWaveletVarModel}
\alias{getWaveletVarModel}
\title{Get fitted model of the fractional Brownian motion wavelet variances}
\usage{
getWaveletVarModel(x)
}
\arguments{
\item{x}{A \code{fracdet} object.}
}
\value{
Either a \code{nls} or a \code{lm} object representing the least-squared
estimates of the fractional Brownian motion parameters.
}
\description{
Get fitted model of the fractional Brownian motion wavelet variances
}
\seealso{
\code{\link{fracdet}}, \code{\link{coef.fracdet}}
}

