% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletVar.R
\name{resolutionLevels}
\alias{resolutionLevels}
\title{Get the resolution levels at which the wavelet coefficients' variances was
computed.}
\usage{
resolutionLevels(x)
}
\arguments{
\item{x}{A \code{WaveletVar} object.}
}
\value{
A numerical vector of the same length as \code{WaveletVar}.
}
\description{
Get the resolution levels at which the wavelet coefficients' variances was
computed.
}

