% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletVar.R
\name{theoreticalWaveletVar}
\alias{theoreticalWaveletVar}
\title{Wavelet coefficients' variances of a fractional Brownian motion}
\usage{
theoreticalWaveletVar(H, sigma2, family, filter_number, nlevels)
}
\arguments{
\item{H}{The Hurst exponent of the fBm.}

\item{sigma2}{Variance of the fractional Gaussian noise that results after
differentiating the fBm.}

\item{family}{Specifies the family of wavelets used in the computation of the
wavelet transform. See \code{\link[wavethresh]{wd}}.}

\item{filter_number}{An integer identifying the concrete filter used within the
\code{family} of wavelets selected. \code{filter_number} is related with the
smoothness of the wavelet. See \code{\link[wavethresh]{wd}}.}

\item{nlevels}{Number of resolution levels to compute. Thus, the resulting
wavelet coefficients' variances are the expected ones from a fBm of length
\eqn{2 ^ n}.}
}
\value{
A \code{WaveletVar} object with the theoretical wavelet variance.
}
\description{
\code{theoreticalWaveletVar} calculates the theoretical wavelet coefficients'
 variance of a fractional Brownian motion (fBm) with known parameters.
}
\examples{
use_H = 0.3
fbm = fbmSim(n = 2 ^ 10, H = use_H)
w_fbm = wd(fbm, bc = "symmetric")
vpr = WaveletVar(w_fbm)
plot(vpr)
# the theoreticalWaveletVar also returns a 'WaveletVar' object:
theo_vpr = theoreticalWaveletVar(H = use_H, sigma2 = 1,
                                 family = wtInfo(vpr)[["family"]],
                                 filter_number = wtInfo(vpr)[["filter_number"]],
                                 nlevels = length(vpr))
points(theo_vpr, col = "red")
}
\seealso{
\code{\link{WaveletVar}}, \code{\link{estimatefBmPars}}
}

