% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveletVar.R
\name{wtInfo}
\alias{wtInfo}
\title{Get the family and the filter-number used in the wavelet transform asociated
with the \code{WaveletVar} object.}
\usage{
wtInfo(x)
}
\arguments{
\item{x}{A \code{WaveletVar} object.}
}
\value{
An R list containing the family (\code{family} field) and the
filter-number (\code{filter_number}).
}
\description{
Get the family and the filter-number used in the wavelet transform asociated
with the \code{WaveletVar} object.
}

