% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracdet.R
\name{fracdet}
\alias{fracdet}
\title{Fractal-deterministic model}
\usage{
fracdet(x, fbmPars)
}
\arguments{
\item{x}{A \code{wd} object representing the wavelet transform of the observed
signal \eqn{Y}.}

\item{fbmPars}{Either a \code{nls} or a \code{lm} object (see
\code{\link[stats]{nls}} and \code{\link[stats]{nls}}) providing estimates
of the parameters characterizing the fBm part of the model (The Hurst
exponent \eqn{H} and the "dispersion" parameter \eqn{\sigma ^ 2}{sigma ^ 2}).
Thus, the names of the fitted variables are assumed to be \code{H} and
\code{sigma2}.}
}
\value{
A S3 \code{fracdet} object that represents the observed \eqn{Y} signal
in the wavelet domain. The \code{fracdet} object will also contain the
estimates of the fBm parameters.
}
\description{
The \code{fracdet} class represents a wavelet transform that is assumed to
be modelled  by a simple fractal-deterministic model with long-memory
properties. The model consists on a  linear superposition of a fractional
brownian motion (fBm) \eqn{B} and a deterministic band-limited signal
\eqn{x}. That is, the observed signal is \eqn{Y = x + B}. The \code{fracdet}
receives as parameters the wavelet transform of \eqn{Y} and an R fitted model
object (\code{nls} or \code{lm}) containing the estimation of the parameters
characterizing the fBm (The Hurst exponent \eqn{H} and the "dispersion"
parameter \eqn{\sigma ^ 2}{sigma ^ 2}). The main method of the \code{fracdet} class
is \code{estimateDetSignal}, which permits the estimation of the deterministic
signal \eqn{x} using bayesian modelling techniques in the wavelet-domain.
}
\details{
In addition to the specific methods for the \code{fracdet} class, all the
\code{wd} methods can be used with a \code{fracdet} object.
}
\note{
When computing the wavelet transform of the \eqn{Y} signal through the
\code{wd} method (see \code{\link[wavethresh]{wd}}) we recommend using the
"symmetric" boundary handling method (\code{bc == "symmetric"}) if
\eqn{H > 0.5}.
}
\examples{
\dontrun{
### deterministic signal estimation
# set parameters for the example
set.seed(3)
nlevels = 14
n = 2 ^ nlevels
xlim = c(200, 400)
H = 0.3
# simulate a simple fbm + sinusal signal
x = 1.25 * cospi(2 * 1:n / 10)
y = fbmSim(n = n, H = H) + x
plot(y, xlim = xlim,
     ylim = range(y[xlim[[1]]:xlim[[2]]]),
     type = "l")
lines(x, lty = 2, col = 2)
legend("topright", lty = 1:2, col = 1:2,
       legend = c("Y", "x"), bty = "n")
# compute the wavelet transform and the wavelet coefficients' variances
wy = wd(y, bc = "symmetric")
vpr = waveletVar(wy)
# do you note the increase in variance in level 11?
plot(vpr, xlim = c(4, nlevels - 1),
     ylim = range(vpr[5:nlevels]))
# estimate the fBm parameters avoiding levels 11 and 12 (with deterministic
# contributions). Level 12 is also avoid as a precaution
model = estimateFbmPars(vpr, use_resolution_levels = c(5:10, 13))

# Create a fracdet object...
fd = fracdet(wy, model)
# ... and check the fbm parameters estimates
coef(fd)
# A plot of the fitted variances may be useful
plot(getWaveletVar(fd))
points(getFittedWaveletVar(fd),
      col = 2,
      pch = 2)
# we could also extract the nls-fit and use the predict function
# The nls-fit is performed in semilog-space. Thus, a transformation
# of the predicted values is required
nls_model = getWaveletVarModel(fd)
points(resolutionLevels(vpr),
       2 ^ predict(nls_model, newdata = data.frame(x = resolutionLevels(vpr))),
       col = 3,
       pch = 3)
# Estimate the deterministic signal (this may take a while) taking
# into account the deviations in level 11, 12. The estimateDetSignal uses
# this information using a Bayesian modelling approach
wx = estimateDetSignal(fd, estimate_from = 11:12)
x_est = wr(wx)
# compare the original signal and the estimation
old_par = par(mfrow = c(2,1))
plot(y, xlim = xlim,
     ylim = range(y[xlim[[1]]:xlim[[2]]]),
     type = "l")
lines(x, lty = 2, col = 2)
legend("topright", lty = 1:2, col = 1:2,
       legend = c("Y", "x"), bty = "n")
plot(x, type = "l", xlim = xlim,
     ylim = range(x) * c(1, 2))
lines(x_est, col = 2, lty = 2)
legend("topright", lty = 1:2, col = 1:2,
       legend = c("x", "x estimate"), bty = "n")
par(old_par)
}
}
\seealso{
\code{\link{waveletVar}}, \code{\link{getWaveletVar}},
\code{\link{getWaveletVarModel}}, \code{\link{getFittedWaveletVar}},
\code{\link{as.wd}}, \code{\link{estimateDetSignal}}
}

