% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracdet.R
\name{getFittedWaveletVar}
\alias{fitted.Fracdet}
\alias{getFittedWaveletVar}
\title{Estimates of the wavelet variances of the fractional Brownian motion}
\usage{
getFittedWaveletVar(x)

\method{fitted}{Fracdet}(object, ...)
}
\arguments{
\item{x, object}{A \code{Fracdet} object.}

\item{...}{Ignored (used for S3 generic/method consistency).}
}
\value{
A \code{WaveletVar} representing the estimates of the wavelet
coefficients' variances of the fractional Brownian motion (fBm) part of the
fractal-deterministic model.
}
\description{
\code{getFittedVar} estimates the wavelet coefficients' variances of the
fractional Brownian motion (fBm) part of the fractal-deterministic model.
The variances estimates are obtained considering the theoretical wavelet
 variance expression and the estimates of the fBm  parameters.
}
\seealso{
\code{\link{Fracdet}}, \code{\link{theoreticalWaveletVar}}
}

