% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveletVar.R
\name{waveletVar}
\alias{waveletVar}
\title{Wavelet coefficients' variances.}
\usage{
waveletVar(x)
}
\arguments{
\item{x}{A \code{wd} object (see \code{\link[wavethresh]{wd}}). The wavelet
coefficients' variances are computed in each of the resolution levels
of \code{wd}.}
}
\value{
A S3 \code{waveletVar} object that stores the
Wavelet coefficients' variances depending on the resolution level.
}
\description{
\code{waveletVar} computes the wavelet coefficients' variances of an object
representing a wavelet transform in each of its resolution levels.
}
\details{
In addition to the specific methods for the \code{waveletVar} class, all the
methods for numeric vectors can be used with a \code{waveletVar} object.
}
\examples{
use_H = 0.3
fbm = fbmSim(n = 2 ^ 10, H = use_H)
w_fbm = wd(fbm, bc = "symmetric")
vpr = waveletVar(w_fbm)
plot(vpr)
# the theoreticalWaveletVar also returns a 'waveletVar' object:
theo_vpr = theoreticalWaveletVar(H = use_H, sigma2 = 1,
                                 family = wtInfo(vpr)[["family"]],
                                 filter_number = wtInfo(vpr)[["filter_number"]],
                                 nlevels = length(vpr))
points(theo_vpr, col = "red")
}
\seealso{
\code{\link{wtInfo}}, \code{\link{resolutionLevels}},
\code{\link{theoreticalWaveletVar}}, \code{\link{fracdet}},
\code{\link{estimateFbmPars}}
}

